feat: small refactor, fix comment
