+++ app-server/src/api/v1/evals.rs
@@ -1,15 +1,15 @@
 use std::{collections::HashMap, sync::Arc};
 
 use crate::{
-    ch::evaluation_datapoints::CHEvaluationDatapoint,
     db::{self, DB, project_api_keys::ProjectApiKey},
     evaluations::{
-        EvaluationDatapointResult, UpdatedDatapointStrings, insert_evaluation_datapoints,
+        EvaluationDatapointResult, UpdatedDatapointStrings,
+        insert_evaluation_datapoints,
+        realtime::{RealtimeDatapoint, publish_datapoint_upsert, publish_inserted_datapoints},
         update_evaluation_datapoint,
     },
     names::NameGenerator,
     pubsub::PubSub,
-    realtime::{SseMessage, send_to_key},
     routes::types::ResponseResult,
 };
 use actix_web::{
@@ -86,19 +86,7 @@ pub async fn save_eval_datapoints(
     )
     .await?;
 
-    if !ch_rows.is_empty() {
-        let realtime_points: Vec<RealtimeDatapoint> = ch_rows
-            .iter()
-            .map(RealtimeDatapoint::from_ch_insert)
-            .collect();
-        publish_datapoint_upsert(
-            pubsub.get_ref().as_ref(),
-            &project_id,
-            &eval_id,
-            &realtime_points,
-        )
-        .await;
-    }
+    publish_inserted_datapoints(pubsub.get_ref().as_ref(), &project_id, &eval_id, &ch_rows).await;
 
     Ok(HttpResponse::Ok().json(eval_id))
 }
@@ -159,143 +147,4 @@ pub async fn update_eval_datapoint(
     .await;
 
     Ok(HttpResponse::Ok().json(datapoint_id))
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-struct RealtimeDatapoint<'a> {
-    id: Uuid,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    index: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    trace_id: Option<Uuid>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    data: Option<&'a str>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    target: Option<&'a str>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    metadata: Option<&'a str>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    output: Option<&'a str>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    scores: Option<&'a str>,
-}
-
-impl<'a> RealtimeDatapoint<'a> {
-    fn from_ch_insert(row: &'a CHEvaluationDatapoint) -> Self {
-        let trace_id = (!row.trace_id.is_nil()).then_some(row.trace_id);
-        Self {
-            id: row.id,
-            index: Some(row.index),
-            trace_id,
-            data: non_empty(&row.data).map(|s| clip_str(s, TRUNCATE_CHARS)),
-            target: non_empty(&row.target).map(|s| clip_str(s, TRUNCATE_CHARS)),
-            metadata: non_empty(&row.metadata),
-            output: non_empty(&row.executor_output).map(|s| clip_str(s, TRUNCATE_CHARS)),
-            scores: non_empty(&row.scores),
-        }
-    }
-
-    fn from_update_strings(
-        datapoint_id: Uuid,
-        trace_id: Option<Uuid>,
-        executor_output: &'a str,
-        scores: &'a str,
-    ) -> Self {
-        let trace_id = trace_id.filter(|t| !t.is_nil());
-        Self {
-            id: datapoint_id,
-            index: None,
-            trace_id,
-            data: None,
-            target: None,
-            metadata: None,
-            output: non_empty(executor_output).map(|s| clip_str(s, TRUNCATE_CHARS)),
-            scores: non_empty(scores),
-        }
-    }
-}
-
-const TRUNCATE_CHARS: usize = 200;
-
-fn clip_str(s: &str, max_chars: usize) -> &str {
-    match s.char_indices().nth(max_chars) {
-        Some((byte_idx, _)) => &s[..byte_idx],
-        None => s,
-    }
-}
-
-fn non_empty(s: &str) -> Option<&str> {
-    match s.trim() {
-        "" | "{}" | "[]" | "null" => None,
-        _ => Some(s),
-    }
-}
-
-async fn publish_datapoint_upsert(
-    pubsub: &PubSub,
-    project_id: &Uuid,
-    evaluation_id: &Uuid,
-    datapoints: &[RealtimeDatapoint<'_>],
-) {
-    if datapoints.is_empty() {
-        return;
-    }
-    let message = SseMessage {
-        event_type: "datapoint_upsert".to_string(),
-        data: serde_json::json!({ "datapoints": datapoints }),
-    };
-    let key = format!("evaluation_{}", evaluation_id);
-    send_to_key(pubsub, project_id, &key, message).await;
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::json;
-
-    #[test]
-    fn clip_str_short_unchanged() {
-        assert_eq!(clip_str("hello", TRUNCATE_CHARS), "hello");
-    }
-
-    #[test]
-    fn clip_str_long_clips_to_max() {
-        let s: String = "a".repeat(500);
-        assert_eq!(clip_str(&s, TRUNCATE_CHARS).chars().count(), TRUNCATE_CHARS);
-    }
-
-    #[test]
-    fn clip_str_multibyte_clips_on_char_boundary() {
-        // 4-byte char × 250 = 250 chars / 1000 bytes. Clip to 200 chars.
-        let s: String = "𝄞".repeat(250);
-        let out = clip_str(&s, TRUNCATE_CHARS);
-        assert_eq!(out.chars().count(), TRUNCATE_CHARS);
-        assert!(out.chars().all(|c| c == '𝄞'));
-    }
-
-    #[test]
-    fn realtime_datapoint_serializes_scores_as_passthrough_string() {
-        let row = CHEvaluationDatapoint {
-            id: Uuid::nil(),
-            evaluation_id: Uuid::nil(),
-            project_id: Uuid::nil(),
-            trace_id: Uuid::nil(),
-            updated_at: 0,
-            data: "{}".into(),
-            target: "{}".into(),
-            metadata: "{}".into(),
-            executor_output: String::new(),
-            index: 0,
-            dataset_id: Uuid::nil(),
-            dataset_datapoint_id: Uuid::nil(),
-            dataset_datapoint_created_at: 0,
-            group_id: "default".into(),
-            scores: r#"{"accuracy":0.9}"#.into(),
-        };
-        let dp = RealtimeDatapoint::from_ch_insert(&row);
-        let v = serde_json::to_value(&dp).unwrap();
-        // Scores arrive as a single string field — frontend handles flattening.
-        assert_eq!(v["scores"], json!(r#"{"accuracy":0.9}"#));
-    }
 }
