+++ app-server/src/evaluations/mod.rs
@@ -1,3 +1,5 @@
+pub mod realtime;
+
 use std::collections::HashMap;
 
 use anyhow::Result;
